// frontend/src/api/models/AppDbModelDocumentVideoProcessingApplication.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const AppDbModelDocumentVideoProcessingApplication = {
    GenerateShorts: 'GENERATE_SHORTS',
    GenerateSubtitling: 'GENERATE_SUBTITLING',
    ApplyTrim: 'APPLY_TRIM',
    ApplySubtitles: 'APPLY_SUBTITLES',
    ApplyVoiceOver: 'APPLY_VOICE_OVER',
    ApplyVlog: 'APPLY_VLOG',
    ApplyPodcastTemplate: 'APPLY_PODCAST_TEMPLATE'
} as const;
export type AppDbModelDocumentVideoProcessingApplication = typeof AppDbModelDocumentVideoProcessingApplication[keyof typeof AppDbModelDocumentVideoProcessingApplication];


export function instanceOfAppDbModelDocumentVideoProcessingApplication(value: any): boolean {
    for (const key in AppDbModelDocumentVideoProcessingApplication) {
        if (Object.prototype.hasOwnProperty.call(AppDbModelDocumentVideoProcessingApplication, key)) {
            if (AppDbModelDocumentVideoProcessingApplication[key as keyof typeof AppDbModelDocumentVideoProcessingApplication] === value) {
                return true;
            }
        }
    }
    return false;
}

export function AppDbModelDocumentVideoProcessingApplicationFromJSON(json: any): AppDbModelDocumentVideoProcessingApplication {
    return AppDbModelDocumentVideoProcessingApplicationFromJSONTyped(json, false);
}

export function AppDbModelDocumentVideoProcessingApplicationFromJSONTyped(json: any, ignoreDiscriminator: boolean): AppDbModelDocumentVideoProcessingApplication {
    return json as AppDbModelDocumentVideoProcessingApplication;
}

export function AppDbModelDocumentVideoProcessingApplicationToJSON(value?: AppDbModelDocumentVideoProcessingApplication | null): any {
    return value as any;
}

export function AppDbModelDocumentVideoProcessingApplicationToJSONTyped(value: any, ignoreDiscriminator: boolean): AppDbModelDocumentVideoProcessingApplication {
    return value as AppDbModelDocumentVideoProcessingApplication;
}